requirements.txt
blinker==1.8.1
click==8.1.7
flask==3.0.3
importlib-metadata==7.1.0
itsdangerous==2.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
PyYAML==6.0.1
werkzeug==3.0.2
zipp==3.18.1
